Add files via upload
This is a responsive e-commerce website that I created using HTML, CSS, and JavaScript. The project showcases the essential features of an online shopping platform, such as product listings with images and prices, a shopping cart system, user login and sign-up pages, and a checkout page design. The main goal of building this project was to practice and enhance my front-end web development skills while creating a simple, user-friendly, and interactive shopping experience.